[security]외부 인증제공자 토큰 기반 로그인을 위한 Authentication process 구현

+++ motherchoice/src/main/java/com/educhoice/motherchoice/configuration/security/service/AccountDetailsService.java
@@ -18,7 +18,7 @@ public class AccountDetailsService implements UserDetailsService {
 		return new SecurityAccount(integratedUserQueryService.loadByEmail(s));
 	}
 
-	public UserDetails loadUserBySocialId(int socialId) throws UsernameNotFoundException {
+	public UserDetails loadUserBySocialId(long socialId) throws UsernameNotFoundException {
 	    return new SecurityAccount(integratedUserQueryService.loadBySocialId(socialId));
     }
