+++ motherchoice/src/main/java/com/educhoice/motherchoice/models/persistent/repositories/AccountRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 public interface AccountRepository extends CrudRepository<Account, Long>{
 
     Optional<Account> findByLoginId(String loginId);
-    Optional<Account> findBySocialId(int id);
+    Optional<Account> findBySocialId(long id);
     Optional<Account> findByNicknameContaining(String nickname);
 }
